Fix TypeScript config: Add JSX support and rename .ts to .tsx
