+++ examples/credit-scoring-app/tsconfig.json
@@ -5,7 +5,8 @@
     "rootDir": "./src",
     "target": "ES2020",
     "module": "commonjs",
-    "lib": ["ES2020"],
+    "lib": ["ES2020", "DOM"],
+    "jsx": "react-jsx",
     "strict": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
